move specs over to container
